+++ Plugins/Regex/Regex.cpp
@@ -32,3 +32,30 @@ NWNX_EXPORT ArgumentStack Replace(ArgumentStack&& args)
 
     return retVal;
 }
+
+NWNX_EXPORT ArgumentStack Match(ArgumentStack&& args)
+{
+    const auto str = args.extract<std::string>();
+      ASSERT_OR_THROW(!str.empty());
+    const auto regex = args.extract<std::string>();
+      ASSERT_OR_THROW(!regex.empty());
+
+    std::regex rgx(regex);
+    std::sregex_iterator it(str.begin(), str.end(), rgx);
+    const static std::sregex_iterator end;
+
+    JsonEngineStructure retVal(json::array());
+
+    while (it != end)
+    {
+        json matches = json::array();
+        for (size_t i = 0; i < it->size(); i++)
+        {
+            matches.emplace_back(it->str(i));
+        }
+        retVal.m_json.emplace_back(matches);
+        it++;
+    }
+
+    return retVal;
+}